Remove redundant log

+++ artifactory/commands/utils/precheckrunner/checkrunner_test.go
@@ -37,12 +37,12 @@ func TestRunChecks(t *testing.T) {
 	for i := uint(0); i < nSuccess; i++ {
 		runner.AddCheck(successCheck)
 	}
-	runAndAssert(t, uint(nSuccess), 0, nil, runner)
+	runAndAssert(t, nSuccess, 0, nil, runner)
 	// With failed checks
 	for i := uint(0); i < nFail; i++ {
 		runner.AddCheck(failCheck)
 	}
-	runAndAssert(t, uint(nSuccess), uint(nFail), nil, runner)
+	runAndAssert(t, nSuccess, uint(nFail), nil, runner)
 	// With check that has error
 	runner.AddCheck(errCheck)
 	runAndAssert(t, 0, 0, expectedErr, runner)
